project/requirements.txt
anthropic
python-dotenv
pandas
numpy
scipy
flask
